cleaned-up gcc compilation

+++ crypt/kovsh/kovsh_deprotect.cpp
@@ -8,7 +8,7 @@
 void pgm_decrypt_dw2(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i<rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -25,7 +25,7 @@ void pgm_decrypt_dw2(std::span<uint8_t> rom)
 void pgm_decrypt_dw3(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -42,7 +42,7 @@ void pgm_decrypt_dw3(std::span<uint8_t> rom)
 void pgm_decrypt_killbld(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -85,7 +85,7 @@ static const uint8_t kov_tab[256] = {
 void pgm_decrypt_kov(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -127,7 +127,7 @@ static const uint8_t kovsh_tab[256] = {
 void pgm_decrypt_kovsh(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -169,7 +169,7 @@ static const uint8_t kovshp_tab[256] = {
 void pgm_decrypt_kovshp(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -211,7 +211,7 @@ static const uint8_t photoy2k_tab[256] = {
 void pgm_decrypt_photoy2k(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -253,7 +253,7 @@ static const uint8_t puzlstar_tab[256] = {
 void pgm_decrypt_puzlstar(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -294,7 +294,7 @@ static const uint8_t puzzli2_tab[256] = {
 void pgm_decrypt_puzzli2(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -336,7 +336,7 @@ static const uint8_t oldsplus_tab[256] = {
 void pgm_decrypt_oldsplus(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -378,7 +378,7 @@ static const uint8_t py2k2_tab[256] = {
 void pgm_decrypt_py2k2(std::span<uint8_t> rom) // and ddp3
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -420,7 +420,7 @@ static const uint8_t ketsui_tab[256] = {
 void pgm_decrypt_ketsui(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -462,7 +462,7 @@ static const uint8_t espgal_tab[256] = {
 void pgm_decrypt_espgaluda(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -485,7 +485,7 @@ void pgm_decrypt_espgaluda(std::span<uint8_t> rom)
 void pgm_decrypt_svg(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -527,7 +527,7 @@ static const uint8_t svgpcb_tab[0x100] = {
 void pgm_decrypt_svgpcb(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -569,7 +569,7 @@ static uint8_t dfront_tab[256] = {
 void pgm_decrypt_dfront(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 	    uint16_t x = (src[i]);
@@ -611,7 +611,7 @@ static uint8_t ddp2_tab[256] = {
 void pgm_decrypt_ddp2(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
     	uint16_t x = (src[i]);
@@ -651,7 +651,7 @@ static uint8_t mm_tab[256] = {
 void pgm_decrypt_martmast(std::span<uint8_t> rom) // and dw2001
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
     	uint16_t x = (src[i]);
@@ -675,7 +675,7 @@ void pgm_decrypt_martmast(std::span<uint8_t> rom) // and dw2001
 void pgm_decrypt_dwpc(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
     	uint16_t x = (src[i]);
@@ -717,7 +717,7 @@ static const uint8_t kov2_tab[256] = {
 void pgm_decrypt_kov2(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
     	uint16_t x = (src[i]);
@@ -758,7 +758,7 @@ static const uint8_t kov2p_tab[256] = {
 void pgm_decrypt_kov2p(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -800,7 +800,7 @@ static const uint8_t theglad_tab[256] = {
 void pgm_decrypt_theglad(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -842,7 +842,7 @@ static const uint8_t killbldp_tab[] = {
 void pgm_decrypt_killbldp(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i< rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -884,7 +884,7 @@ static const uint8_t happy6in1_tab[256] = {
 void pgm_decrypt_happy6(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
@@ -926,7 +926,7 @@ static const uint8_t pgm3in1_tab[256] = {
 void pgm_decrypt_pgm3in1(std::span<uint8_t> rom)
 {
 	uint16_t *src = (uint16_t *)rom.data();
-	int const rom_size = rom.size();
+	size_t const rom_size = rom.size();
 
 	for (uint32_t i = 0; i < rom_size/2; i++) {
 		uint16_t x = (src[i]);
